Some refactoring to handle QCs better
